refactor: remove fastapi app factory from common
The app factory abstraction has been removed from the common package.
The app factory was a way to create a FastAPI app with some common
middleware and configuration. This was a good idea in theory, but
ultimately it was not flexible enough to support the needs of the
services. The app factory was removed and the common package was
simplified to make it easier for services to use its logging config.